VirtualBox guest packages update
Summary:
- Drop additions_fix_4.17.patch as it is not required anymore
- Stop building kernel modules for current as they have been enabled at kernel level
- Current : build `vboxsf` module from [Hans de Goede repository](https://github.com/jwrdegoede/vboxsf) (until it's mainlined in the kernel)
- Move `mount.vboxsf` from common to lts package as current can now use the regular mount binary

Automount now works again fine on both `current` and `lts`.
Using mainline kernel modules will offer a better out-of-box experience when using Solus in a virtual machine.

Depends on D3695

Signed-off-by: Pierre-Yves <pyu@riseup.net>

Test Plan: Tested Solus on virtual machines with the `current` and with the `lts` kernels and the corresponding guest packages

Reviewers: #triage_team, DataDrake

Reviewed By: #triage_team, DataDrake

Subscribers: DataDrake

Differential Revision: https://dev.solus-project.com/D3696